resolution des exercices

+++ basics/src/main/java/be/ecam/basics/exercises/SensorWindow.java
@@ -1,11 +1,11 @@
 package be.ecam.basics.exercises;
 
 public class SensorWindow {
-    public static int average(int[] values) {
+    public static long average(int[] values) {
         if (values == null || values.length == 0) throw new IllegalArgumentException("values");
-        int sum = 0;
+        long sum = 0;
         for (int v : values) {
-            sum += v;
+            sum += (long)v;
         }
         return sum / values.length;
     }
